0414 v2
add function update_price()

necessary changes to other functions

+++ print_market.h
@@ -2,5 +2,5 @@
 #ifndef _PRINT_MARKET_H
 #define _PRINT_MARKET_H
 
-void print_market(Asset a, int r);
+void print_market(Asset a, double r);
 #endif
